Fold AI-accounts management into the subrouter dashboard section
PR #7330 shipped /dashboard/ai-accounts (team AI provider account
management) in a different visual language than the rest of the
dashboard, rendered its own marketing SiteHeader inside the dashboard
shell (double chrome), and left it unlinked from the shell nav while
/dashboard/subrouter was a separate "coming soon" stub.

Moves the full account-management UI (team switcher, accounts list, add
Claude/Anthropic/Codex/OpenAI forms, delete, notConfigured/error states)
into /dashboard/subrouter, restyled to the dashboard design language:
square corners, monochrome tokens only, two font sizes, plain team-
switcher links, 1px-bordered sections, mono created-at dates, invert-on-
hover bordered buttons, and the standard container + getUser return-null
+ vaultSignInHref guard. The marketing SiteHeader/min-h-screen wrapper is
gone (the shell provides chrome). /dashboard/ai-accounts is now a
redirect to /dashboard/subrouter preserving ?team. All subrouter service
and /api/subrouter/accounts behavior is unchanged; only presentation and
route placement moved. Removes the now-unused dashboard.subrouter
comingSoon* keys from en+ja.
