+++ web/messages/en.json
@@ -65,9 +65,7 @@
     },
     "subrouter": {
       "title": "subrouter",
-      "description": "subrouter routes llm requests across provider accounts with failover and usage-aware scheduling.",
-      "comingSoonTitle": "dashboard integration is coming soon",
-      "comingSoonBody": "metrics and controls will appear here after the backend is wired."
+      "description": "subrouter routes llm requests across provider accounts with failover and usage-aware scheduling."
     },
     "aiAccounts": {
       "metaTitle": "AI accounts — cmux",
